feat(emissions): remove redundant app/emissions directory
- Delete entire app/emissions/ directory (278 lines removed)
- All logic successfully moved to x/emissions/types/
- Chain continues running with consolidated emissions

Step 4.1 complete - redundant app layer code cleaned up
